Move tables related code to its own file.
